feat(app): add server-side rendering and database integration
