+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@prisma/client": "^5.0.0",
     "autoprefixer": "10.4.14",
+    "axios": "^1.4.0",
     "eslint": "8.45.0",
     "eslint-config-next": "13.4.12",
     "next": "13.4.12",
